PE-172: combine SchemaUtil into XsdHelper
