--- features/config/dao/impl/src/main/java/org/opennms/features/config/dao/impl/util/SchemaUtil.java
@@ -1,53 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2019-2021 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2021 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- ******************************************************************************/
-
-package org.opennms.features.config.dao.impl.util;
-
-import java.io.IOException;
-import java.net.URL;
-
-import org.springframework.core.io.Resource;
-import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
-
-public class SchemaUtil {
-    /**
-     * It will search xsds first, otherwise it will search across classpath
-     *
-     * @return URL of the xsd file
-     */
-    public static URL getSchemaPath(String xsdName) throws IOException {
-        URL url = Thread.currentThread().getContextClassLoader().getResource("xsds/" + xsdName);
-        if (url == null) {
-            PathMatchingResourcePatternResolver resolver = new PathMatchingResourcePatternResolver();
-            Resource[] resources = resolver.getResources("classpath*:**/" + xsdName);
-            if (resources != null && resources.length > 0)
-                url = resources[0].getURL();
-        }
-        return url;
-    }
-}
