+++ features/config/test/src/test/java/org/opennms/features/config/dao/impl/XsdModelConverterTest.java
@@ -32,8 +32,8 @@ import com.google.common.io.Resources;
 import org.junit.Assert;
 import org.junit.Test;
 import org.opennms.features.config.dao.api.ConfigItem;
-import org.opennms.features.config.dao.impl.util.SchemaUtil;
 import org.opennms.features.config.dao.impl.util.JaxbXmlConverter;
+import org.opennms.features.config.dao.impl.util.XsdHelper;
 import org.opennms.features.config.dao.impl.util.XsdModelConverter;
 
 import java.io.IOException;
@@ -47,7 +47,7 @@ public class XsdModelConverterTest {
         // register
         JaxbXmlConverter converter = new JaxbXmlConverter("provisiond-configuration.xsd", "provisiond-configuration", null);
 
-        String xsdStr = Resources.toString(SchemaUtil.getSchemaPath("provisiond-configuration.xsd"), StandardCharsets.UTF_8);
+        String xsdStr = Resources.toString(XsdHelper.getSchemaPath("provisiond-configuration.xsd"), StandardCharsets.UTF_8);
         XsdModelConverter xsdConverter = new XsdModelConverter(xsdStr);
         ConfigItem item = xsdConverter.convert("provisiond-configuration");
